Debugging pipo protocol

+++ src/parsers/adminRequestParser/adminResponseBuilder.c
@@ -63,7 +63,7 @@ bool admin_response_builder_uint32(AdminResponseBuilderContainer * adminResponse
         }
         else {
             (*currByte)++;
-            aux = (adminResponse->data.uint16 >> ((UINT32_RESPONSE_SIZE - *currByte)* 8)) & MASK;
+            aux = (adminResponse->data.uint32 >> ((UINT32_RESPONSE_SIZE - *currByte)* 8)) & MASK;
             buffer_write(b, aux);
         }
     }
@@ -88,7 +88,7 @@ bool admin_response_builder_uint64(AdminResponseBuilderContainer * adminResponse
         }
         else {
             (*currByte)++;
-            aux = adminResponse->data.uint16 >> ((UINT64_RESPONSE_SIZE - *currByte)* 8) & MASK;
+            aux = adminResponse->data.uint64 >> ((UINT64_RESPONSE_SIZE - *currByte)* 8) & MASK;
             buffer_write(b, aux);
         }
     }
